requirements.txt
anyio==3.6.2
autopep8==2.0.1
certifi==2022.12.7
h11==0.14.0
httpcore==0.16.3
httpx==0.23.2
idna==3.4
pycodestyle==2.10.0
python-dotenv==0.21.0
python-telegram-bot==20.0
rfc3986==1.5.0
sniffio==1.3.0
tomli==2.0.1
